adding Megha's notes - includes commented code and annotated lecture slide deck (#7)
* Add files via upload

This is the code from the live coding demo that happened during lecture 7.
It is the same code as in the discord, but with my additional comments on each line that basically tries to match it up to the algorithm

* lightly annotated slide deck

These are some of the slide decks we have gone through so far with light annotations. 
My notes are in blue font or are little drawings throughout the slide deck.

* heavily annotated slide desk

These are some of the slide decks we have gone through so far with heavier annotations. 
My notes are in blue font or are little drawings throughout the slide deck.